refactor: remove duplicate code and reause extensions.
